Completata implementazione dell'algoritmo Nearest Neighbor. Completata prima implementazione dell'algoritmo Two-Opt con relativo inizio di debug. Sistemati files XML dei punti d'interesse. Bug fixing nell'algoritmo Nearest Neighbor per la selezione del nodo casuale. Refactoring classi: Problem, TOProblem, TOSolution, Result, ResultType, Algorithm.

+++ CityScover.Engine/Problem.cs
@@ -3,7 +3,7 @@
 // Version 1.0
 //
 // Authors: Andrea Ritondale, Andrea Mingardo
-// File update: 16/09/2018
+// File update: 01/10/2018
 //
 
 using System;
@@ -28,9 +28,9 @@ namespace CityScover.Engine
       /// <summary>
       /// Abstract automatic properties doesn't create a private backing field.
       /// </summary>
-      internal abstract Func<TOSolution, double> ObjectiveFunc { get; set; }
+      internal abstract Func<TOSolution, int> ObjectiveFunc { get; set; }
 
-      internal abstract Func<TOSolution, double> PenaltyFunc { get; set; }
+      internal abstract Func<TOSolution, int> PenaltyFunc { get; set; }
       #endregion
 
       #region Internal properties
